debug writer

+++ file_handling_lab/employees.py
@@ -5,7 +5,7 @@ class Employee:
         self.first_name = first_name
         self.last_name = last_name
         self.hourly_rate = float(hourly_rate)
-        self.hours_worked = int(hours_worked)
+        self.hours_worked = float(hours_worked)
         self.amount_due = float(amount_due)
     
     def values(self):
@@ -25,3 +25,14 @@ with open("employees.csv", "a") as csvfile:
     writer = csv.writer(csvfile, quoting=csv.QUOTE_ALL)
     employee = Employee("Jenny","Jones",12.50,40,0)
     writer.writerow(employee.values())
+
+# with open("employees.csv", "w") as csvfile:
+#     writer = csv.writer(csvfile, quoting=csv.QUOTE_ALL)
+
+# # Write out our headers
+#     writer.writerow(["first_name", "last_name", "hourly_rate", "hours_worked", "amount_due"])
+
+#     for employee in employees:
+#         amount_due = (1 * 2)
+#         new_employee = Employee(employee.first_name, employee.last_name, employee.hourly_rate, employee.hours_worked, amount_due)
+#         writer.writerow(new_employee)
